Fixed SENTRY_GUN object not tracking Lara at some angles

+++ TombEngine/Objects/TR4/Entity/tr4_sentry_gun.cpp
@@ -103,7 +103,7 @@ namespace TEN::Entities::TR4
 
 				item.Pose.Position.y += CLICK(2);
 
-				int deltaAngle = ai.angle - creature->JointRotation[0];
+				short deltaAngle = ai.angle - creature->JointRotation[0];
 
 				ai.ahead = true;
 				if (deltaAngle <= ANGLE(-90.0f) || deltaAngle >= ANGLE(90.0f))
